Add utility method print_all_logs (#2)
Other LoansBot repos use this module for the majority of their logging, but that means integration tests are harder to debug. This adds a utility class "query.py" which can be executed to print all logs at the end of an integration test suite.

Most likely dedicated functions by the web backend will be used for production logs anyway, so this isn't very feature rich.